feat: refactored base theme
This PR decreases the amount of Bootstrap customization, by
cutting down the amount of variables. It defaults to Bootstrap
default font and navbar customization, which makes themes more
consistent and easier to maintain. The main variables only
has variables that are meant to be exposed to themes. Docs are
updated and the SCSS service has been fixed.

BREAKING CHANGE: Removed navbar customization; themes will look different